requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
redis
hiredis
requests
authlib
python-dotenv
gunicorn
itsdangerous